[sui-genesis-builder] Start genesis with correct framework for protocol version (#13885)
Previously we would always start genesis with the current set of
framework modules. This causes issues in the compatibility tests if the
current version of the framework is not backwards compatible with the
old version (e.g., by adding a new type in the new version of the
framework).

This updates the genesis creation to now use the correct framework
snapshot for all protocol
versions, defaulting to the current framework if a snapshot for the
given protocol version is not found.

This also updates framework snapshots to be ordered based on the order
in which they need to be published.

## Test Plan 

Added a type to the framework locally, and made sure the
`test_upgrade_compatibility` test passes with these changes.