[graphql/rpc] Impl custom query limiter for early termination (#13996)
## Description 

Current limit checking logic in async-graphql library does not terminate
early if limits are reached. This is wasteful as it allows other
validations and checks to go through.
Created an upstream
[issue](https://github.com/async-graphql/async-graphql/issues/1383) to
support this but in meantime we can run an extension early which
approximates the desired behavior.

In future we can add limits for other query components in this extension

Todo: reconcile this logic with the remaining query logic and extensions

## Test Plan 

Manual
Unit TBD

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes