[narwhal] Add CertificateV2 (#13777)
Added new version for NW `Certificate` which Includes new
`AggregateSignatureState`. This holds `AggregateSignatureBytes` but with
the added layer to specify if the signature was verified via a leader,
verified directly, unverified or unsigned. This will be used to take
advantage of the certificate chain that is formed via the DAG by only
verifying the leaders of the certificate chain when they are fetched
from validators during catchup.

This is gated by a protocol feature flag so this PR should have no
effect yet. Still testing the [followup
PR](https://github.com/MystenLabs/sui/commit/ef48f56d972e46ada7b58bd2ecd7eebd23d8010b)
which will include the usage of `CertificateV2`, but sharing here to
provide more context for these changes.