[graphql]suins lookup (#14197)
## Description 

implement resolveNameServiceAddress on top-level Query resolver +
nameServiceConnection for objectOwner.

Not keen on importing all of sui-json-rpc just to get access to the
NameService type. Made a copy in sui-graphql-rpc, but thoughts on doing
something like lifting the type to sui-json-rpc-types?
<img width="965" alt="Screenshot 2023-10-10 at 3 47 22 PM"
src="https://github.com/MystenLabs/sui/assets/127570466/01ffaf21-dba5-493f-a5dd-8f637af24c15">

Not clear on what defaultNameServiceName should be ... is this intended
to be the first/ primary name returned by
https://docs.sui.io/sui-jsonrpc#suix_resolveNameServiceNames



## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes