[graphql] split queries into build and execution (#14442)
## Description 

Simple refactor that splits querying into two parts, building the query
and actually executing it. This is done to support making an explain
call on the query fragment before every query is actually executed

## Test Plan 

Manually ran all of our examples

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes

---------

Co-authored-by: oxade <93547199+oxade@users.noreply.github.com>