[move compiler] Report unnecessary mutable references (#14216)
## Description 

- Report unused mutable references
- The system is quite exhaustive and reports even locally created
mutable references. It piggy-backs on the borrow checker but globally
collects usages instead of keeping it in the abstract state. This is
done mostly for simplicity
- A mutable reference is considered unused if:
  - It is mutated
- It is passed to a function (since freeze removes any subtyping issues)
  - It is returned 
- Copies share the same abstract reference ID

## Test Plan 

- Added new tests

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [X] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes

With this change, Move will now suggest switching mutable references
`&mut` to immutable ones `&` when the reference is not modified, or
passed somewhere where the mutability is required.
This can be suppressed for parameters by prefixing the parameter name
with an underscore `_`.