[move-prover] Print warning about the level of Prover support for Sui (#14348)
## Description 

People keep trying to use Move Prover with Sui and there is currently no
easily accessible place where the current level of support is mentioned.
Adding a warning at the time when Prover is ran seems like a reasonable
thing to do here.

## Test Plan 

Verified that the warning is printed.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [x] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes
A warning about current level of Move Prover support for Sui is now
printed when the Prover is ran.